Complete refactoring; Splitted project into two; moved AnswerQuestion cycle to main program
