+++ BertCs/BertCs.csproj
@@ -1,9 +1,8 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <PackageId>Knorre.ConsoleBart</PackageId>
-    <Version>0.1.2709.231</Version>
-    <OutputType>Exe</OutputType>
+    <PackageId>Knorre.Bart</PackageId>
+    <Version>0.1.2909.162</Version>
     <TargetFramework>net7.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
